Move validation checks to new file, helper functions to other util files
